feat: Update invoice management system with new features and improvements
- Refactored InvoiceForm to remove tax calculations and added invoice status selection.
- Enhanced InvoicePreview to utilize a utility function for currency formatting.
- Updated PageView to display invoice summary and improved layout for better user experience.
- Modified columns definition for invoices to reflect new data structure and added formatting for total amounts.
- Implemented utility function for currency conversion.
- Added server functions to retrieve invoice revenue and summary.
- Created a new page for editing invoices with proper data mapping and PDF generation functionality.
- Introduced animation transitions for page navigation using framer-motion.

+++ serverfunctions/invoice.ts
@@ -1,5 +1,6 @@
 "use server"
 import { prisma } from "@/lib/prismaClient";
+import { get } from "http";
 import { Inter } from "next/font/google";
 
 
@@ -137,6 +138,15 @@ export async function getInvoiceById(id: string) {
 
 export async function deleteInvoice(id: string) {
     try {
+        const getInvoice = await prisma.invoice.findUnique({
+            where: { id },
+        });
+        if (!getInvoice) {
+            throw new Error("Invoice not found");
+        }
+        if(getInvoice.status === "PAID"){
+            throw new Error("Cannot delete a paid invoice");
+        }
         const deletedInvoice = await prisma.$transaction(async (tx) => {
             // Delete invoice items
             await tx.invoiceItem.deleteMany({
@@ -257,6 +267,36 @@ export async function getInvoicesCount() {
         return count;
     } catch (error) {
         console.error("Error counting invoices:", error);
+        throw error;
+    }
+}
+
+export async function getInvoicesRevenue() {
+    try {
+        const sum = await prisma.invoice.aggregate({
+            _sum:{
+                total:true  
+            }
+        });
+        return sum._sum.total || 0;
+    } catch (error) {
+        console.error("Error counting invoices:", error);
+        throw error;
+    }
+}
+
+
+export async function getInvoicesSummary() {
+    try {
+        const summary = await prisma.invoice.groupBy({
+            by:['status'],
+            _sum:{
+                total:true  
+            }
+        });
+        return summary || [];
+    } catch (error) {
+        console.error("Error getting summary invoices:", error);
         throw error;
     }
 }
